constructors: Impl New ctor func & make adaptations
Implement the New constructor function, making adaptations in the
newCallStack function for avoiding that the New call appears in the
stack.

Also make a refactoring on the error test for extracting common helpers
used across different tests files and use the New constructor on the
derror test which previously used directly the struct type.